Changed Posts backend
Added a new "last_edited" field to the post model, which stores when the post was last edited. It is initially empty, and if a post is edited, the field will be changed to whenever the post was edited.

Additionally, the tests related to editing posts were changed. A new test was made for checking if an edit post request was made without a last_edited field.